Start to improve DDB Local compatability
I have over-asserted on the error messages, so need to laxen those.
Also, it looks like the AWS SDK---or at least, the DynamoDB part of
it---never returns (nil, err); the first argument is a non-nil pointer
to a blank instance of the expected result type. So drop those
assertions since they don't match reality.